backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
django-filter==23.3
drf-yasg==1.21.7
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3.post1
geopy==2.4.0
gunicorn==21.2.0
whitenoise==6.5.0 